+++ server/mailer/teardown.js
@@ -1,0 +1,16 @@
+const mailer = require('./index');
+
+const templateId = 'd-9e472f554acf47f4a626112db0d920ce';
+
+module.exports = function (postId, teardownHash, name, email) {
+    return mailer.send({
+        to: email,
+        from: 'yourfriends@kindnessproject.xyz',
+        templateId,
+        dynamic_template_data: {
+            postId,
+            teardownHash,
+            name,
+        },
+    });
+};